fix icon and many more

+++ laravel/database/migrations/2025_05_01_235801_create_banner_contents.php
@@ -13,7 +13,7 @@ return new class extends Migration {
         Schema::create('banner_contents', function (Blueprint $table) {
             $table->ulid('id')->primary();
             $table->foreignUlid('banner_category_id')->nullable()->index()->nullOnDelete();
-            $table->unsignedSmallInteger('sort')->default(1)->index();
+            $table->unsignedInteger('sort')->default(1)->index();
             $table->boolean('is_active')->default(false)->index();
 
             $table->string('title', 255)->nullable()->index();
